VLM-Bot/requirements.txt
# ============================================================================
# VLM-Bot Requirements - Python 3.10+
# Compatible avec: NVIDIA RTX 4050 Laptop (6GB VRAM, CUDA 12.8)
# Utilise LLaVA-1.5-7B avec quantization 4-bit
# ============================================================================

# ⚠️ IMPORTANT: Installer PyTorch avec CUDA AVANT:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Deep Learning & Transformers
# torch>=2.1.0  # Déjà installé via venv_gpu
transformers>=4.37.0
accelerate>=0.26.0
bitsandbytes>=0.42.0  # Quantization 4-bit pour 6GB VRAM
peft>=0.8.0
sentencepiece>=0.1.99
protobuf>=3.20.0
safetensors>=0.4.1

# Vision & Image Processing
Pillow>=10.1.0

# RAG & Embeddings
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data Processing
datasets>=2.16.0
pandas>=2.1.4
numpy>=1.24.0,<2.0.0  # Limiter pour compatibilité
scikit-learn>=1.3.2

# Color Processing
webcolors>=1.13

# Web Interface
gradio>=4.13.0
gradio-client>=0.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tqdm>=4.66.1
huggingface-hub>=0.20.0
requests>=2.31.0

# Optional: Performance monitoring
psutil>=5.9.0  # Monitoring RAM/GPU
gputil>=1.4.0  # Monitoring GPU usage
